Make obfs4 client save a random seed to disk.
Obfs4 clients are stateless; only obfs4 servers save their configuration
to disk.  This patch makes obfs4 clients generate a random seed upon
first run and save it to disk for subsequent use.  We are currently not
using this seed (clients derive their probability distributions from the
server's public key) but we may decide to use it in the future.